restructered controller section
